refactor: переименует компонент RadioInput и подключит компонент CheckBox RTI-2194
